requirements.txt
appnope==0.1.0
backcall==0.1.0
bcolz==1.2.1
bleach==3.0.2
cached-property==1.5.1
CausalInference==0.1.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
entrypoints==0.2.3
fastprogress==0.1.15
Flask==1.0.2
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
linearmodels==4.11
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
regex==2018.1.10
requests==2.20.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.11.0
spacy==2.0.16
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
ujson==1.35
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
